refactor: project structure for better libraries
