+++ db/connection.go
@@ -1,4 +1,4 @@
-package pkg
+package db
 
 import (
 	"context"
