+++ models/Payload.go
@@ -1,0 +1,5 @@
+package models
+
+type Payload interface {
+	Run(args map[string]interface{}) error
+}
