--- pkg/scheduler.go
@@ -1,9 +1,0 @@
-package pkg
-
-import "github.com/robfig/cron"
-
-// Scheduler -
-type Scheduler struct {
-	ScriptDirectory string
-	Cron            *cron.Cron
-}
