+++ tasker.go
@@ -5,7 +5,9 @@ import (
 
 	"github.com/gorilla/mux"
 	cron "github.com/robfig/cron/v3"
-	"github.com/roger-king/tasker/pkg"
+	"github.com/roger-king/tasker/db"
+	"github.com/roger-king/tasker/handlers"
+	"github.com/roger-king/tasker/services"
 	log "github.com/sirupsen/logrus"
 	"go.mongodb.org/mongo-driver/mongo"
 )
@@ -34,7 +36,7 @@ func init() {
 
 // New - Creates a new instance of tasker
 func New(tc *TaskerConfig) *Tasker {
-	m, err := pkg.NewMongoConnection(tc.MongoConnectionURL)
+	m, err := db.NewMongoConnection(tc.MongoConnectionURL)
 
 	if err != nil {
 		log.Panic(err)
@@ -51,23 +53,11 @@ func (t *Tasker) Start() *mux.Router {
 	log.Info("Starting Tasker application")
 	t.Scheduler.Start()
 
-	taskService := &pkg.TaskService{
+	taskService := &services.TaskService{
 		DB:        t.DB,
 		Scheduler: t.Scheduler,
 	}
 
-	r := mux.NewRouter()
-	apiRouter := r.PathPrefix("/tasker").Subrouter()
-	apiRouter.HandleFunc("/tasks", pkg.ListTasks(taskService)).Methods("GET")
-	apiRouter.HandleFunc("/tasks", pkg.CreateTask(taskService)).Methods("POST")
-
-	// Single Task Routes
-	apiRouter.HandleFunc("/tasks/{taskID}", pkg.FindTask(taskService)).Methods("GET")
-	apiRouter.HandleFunc("/tasks/{taskID}/disable", pkg.DisableTask(taskService)).Methods("PATCH")
-	apiRouter.HandleFunc("/tasks/{taskID}", pkg.DeleteTask(taskService)).Methods("DELETE")
-
-	// Web Admin - We have a reverse proxy for working on local developer :)
-	r.PathPrefix("/static/").HandlerFunc(pkg.ServeWebAdmin)
-	apiRouter.PathPrefix("/admin").HandlerFunc(pkg.ServeWebAdmin)
+	r := handlers.NewRouter(taskService)
 	return r
 }
