+++ utils/constants.go
@@ -1,4 +1,4 @@
-package pkg
+package utils
 
 import (
 	"os"
@@ -25,4 +25,18 @@ var (
 
 	// TASKER_ENV -
 	TaskerEnv = os.Getenv("TASKER_ENV")
+)
+
+// HTTPError -
+type HTTPError string
+
+func (h HTTPError) String() string {
+	return string(h)
+}
+
+const (
+	// ProcessingError -
+	ProcessingError HTTPError = "processing_error"
+	// RequestError -
+	RequestError HTTPError = "request_error"
 )
